+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.14.0
 
+- Refactor `BBadge` and remove `BChip`
 - `BList` refactoring
 
 ## v0.13.0